+++ CHANGES.md
@@ -4,6 +4,7 @@ All notable changes to this project will be documented in this file.
 ## 0.11.0
 - Update load portfolio improve load time 
 - First land implementation
+- Add last investment marker for visual effect.
 
 ## 0.10.0
 - Add monitor of staked DEC